add cross-platform declaration for IsContainerized()
